Fix L1 cache size bug in GUI app
This code change fixes the following rare bug:

The primesieve GUI application for Windows fails to set the default sieve size in the drop down list if the CPU's L1 cache size is not a power of 2. The only CPUs for Windows currently on the market that do not have a power of 2 L1 cache size are currently Intel Atom CPUs.