Make get_sieve_size() use the L2 cache size even when it isn't private
If the L2 is shared we currently use the L1 cache size. With the new
topology changes we know the number of threads per L2, so we can
instead divide the L2 size by the number of threads.

With this change primesieve 1e19 --dist=1e11 is 20% faster on POWER9 in ST
mode.