Speeds up PrimeGenerator::fill (proof-of-concept).
On my laptop, this commit reduces the #cycles used by this
PrimeGenerator::fill by about 45%, when used via primesieve::iterator
to generate all primes up to 50 billion. Tested with clang only.

At this point, this is a proof-of-concept only. If this direction is
aligned with the goals of primesieve, this code would have to be
cleaned and made portable.

There are a number of negative side effects of this approach. The code
is more complex than the original. The speed up relies heavily on
branch prediction, out-of-order execution, and POPCNT, so it may
actually be a regression on other processors (I did not test that).

Also, it seems that this code slows down multi-threaded functionality
such as NthPrime, PrimePi etc. So primesieve --test is slower.

Stats on Intel Skylake, for generating all primes up to 50 billion.

  primesieve::iterator it;
  __int128_t sum = 0;
  for (int64_t p = it.next_prime(); p <= 50'000'000'000; p = it.next_prime())
    sum += p;
  cout << sum << endl;

Compiled with clang 11 -std=c++2a -O3 -pedantic -fomit-frame-pointer -m64 -march=skylake.

Original code:
47,975,954,230      cycles                    #    4.185 GHz
94,568,666,674      instructions              #    1.97  insn per cycle
13,884,302,579      branches                  # 1211.245 M/sec
   561,589,446      branch-misses             #    4.04% of all branches
        23.89%      %cycles in PrimeGenerator::fill

New code:
42,332,943,051      cycles                    #    4.154 GHz
92,384,340,392      instructions              #    2.18  insn per cycle
11,974,220,227      branches                  # 1174.925 M/sec
   341,654,578      branch-misses             #    2.85% of all branches
        14.62%      %cycles in PrimeGenerator::fill